The options "Fight" makes sense. The pokemon, run and bag do not. We need to rethink those to apply to the game. Lets avoid this instead and simply have FIGHT, Instructions, and Reset Game options please.
And the video camera should not overlap with the arena pieces. Lets add some animation to open up space for it. Or have it as a simple overlay where the entire screen has an overlay and visible count down with 3 2 1 in white font where the numbers are as big as the screen before taking the shot. Once thats done, it should go immediately to the result against the computer / gameplay opponent.

Also please add a white polka-dot background. And add a different font, a fun one that still caters to the 20-30 yr old audience.