Lets make add a new feature to this reto rumble. Lets add another CTA called "Shoot!" where it show a visual 3 2 1 countdown before taking a screenshot of the user (must ask user for permission using their camera before CTA can be visible). Once the screenshot is done, conclude the gameplay and show the end result against the computer.
Also, each game should lower the users health by a certain percentage. I suppose it should be best of 3. So, 33% at a time should be taken each game until there is a winner.Once the winner is determined throw a celebration confetti animation  at the end and declare the victor.